Revert the introduction of SlotArray container
Memory management should be simple. I made a generalized class that in the end does more than it should and optimized prematurely. Stick to STL containers and optimize when things are actually slow. Vectors are slow for bigger types, but for small ones they arent so bad to reallocate even for a milion of elements.